+++ crates/pathfinder/src/bin/pathfinder/config.rs
@@ -3,12 +3,12 @@ use std::collections::{BTreeMap, HashMap, HashSet};
 use std::fs::File;
 use std::net::SocketAddr;
 use std::num::{NonZeroU32, NonZeroU64, NonZeroUsize};
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use std::time::Duration;
 
 use clap::{ArgAction, CommandFactory, Parser};
 use pathfinder_common::{AllowedOrigins, StarknetVersion};
-use pathfinder_executor::VersionedConstantsMap;
+use pathfinder_executor::{VersionedConstants, VersionedConstantsMap};
 use pathfinder_storage::JournalMode;
 use reqwest::Url;
 use util::percentage::Percentage;
@@ -636,10 +636,10 @@ enum RpcCorsDomainsParseError {
 }
 
 fn parse_versioned_constants(
-    path: PathBuf,
+    path: &Path,
 ) -> Result<VersionedConstantsMap, ParseVersionedConstantsError> {
     let mut target = BTreeMap::new();
-    let file = File::open(path.clone())?;
+    let file = File::open(path)?;
     let reader = std::io::BufReader::new(file);
     let src_res: Result<HashMap<String, String>, _> = serde_json::from_reader(reader);
     if let Ok(source) = src_res {
@@ -650,23 +650,19 @@ fn parse_versioned_constants(
         })?;
         for (raw_version, rel_path) in source {
             let version = raw_version.parse::<StarknetVersion>().map_err(|_| {
-                ParseVersionedConstantsError::Parse(serde::de::Error::custom(format!(
+                ParseVersionedConstantsError::ParseMap(serde::de::Error::custom(format!(
                     "Invalid Starknet version \"{}\"",
                     raw_version
                 )))
             })?;
             let abs_path = std::fs::canonicalize(dir_path.join(rel_path))?;
-            let file = File::open(abs_path)?;
-            let reader = std::io::BufReader::new(file);
-            let constants = serde_json::from_reader(reader)?;
+            let constants = VersionedConstants::from_path(&abs_path)?;
             target.insert(version, Cow::Owned(constants));
         }
     } else {
         // logging isn't set up yet...
         eprintln!("Unknown versioned constants map file format - trying legacy...");
-        let file = File::open(path)?;
-        let reader = std::io::BufReader::new(file);
-        let constants = serde_json::from_reader(reader)?;
+        let constants = VersionedConstants::from_path(path)?;
         target.insert(
             VersionedConstantsMap::latest_version(),
             Cow::Owned(constants),
@@ -674,7 +670,7 @@ fn parse_versioned_constants(
     }
 
     if target.is_empty() {
-        return Err(ParseVersionedConstantsError::Parse(
+        return Err(ParseVersionedConstantsError::ParseMap(
             serde::de::Error::custom("Version constants map file specified but empty"),
         ));
     }
@@ -682,7 +678,7 @@ fn parse_versioned_constants(
     Ok(VersionedConstantsMap::custom(target))
 }
 
-pub fn parse_versioned_constants_or_exit(path: PathBuf) -> VersionedConstantsMap {
+pub fn parse_versioned_constants_or_exit(path: &Path) -> VersionedConstantsMap {
     use clap::error::ErrorKind;
 
     match parse_versioned_constants(path) {
@@ -697,8 +693,10 @@ pub fn parse_versioned_constants_or_exit(path: PathBuf) -> VersionedConstantsMap
 enum ParseVersionedConstantsError {
     #[error("IO error while reading versioned constants: {0}.")]
     Io(#[from] std::io::Error),
+    #[error("Parse error while loading versioned constants map: {0}.")]
+    ParseMap(#[from] serde_json::Error),
     #[error("Parse error while loading versioned constants: {0}.")]
-    Parse(#[from] serde_json::Error),
+    Parse(#[from] pathfinder_executor::VersionedConstantsError),
 }
 
 pub struct Config {
@@ -922,7 +920,7 @@ impl Config {
             state_tries: cli.state_tries,
             versioned_constants_map: cli
                 .custom_versioned_constants_path
-                .map(parse_versioned_constants_or_exit)
+                .map(|path| parse_versioned_constants_or_exit(&path))
                 .unwrap_or_default(),
             fetch_casm_from_fgw: cli.fetch_casm_from_fgw,
             shutdown_grace_period: Duration::from_secs(cli.shutdown_grace_period.get()),
@@ -1052,7 +1050,7 @@ mod tests {
     #[test]
     fn parse_versioned_constants_fails_if_file_not_found() {
         assert_matches!(
-            super::parse_versioned_constants("./nonexistent_versioned_constants.json".into()).unwrap_err(),
+            super::parse_versioned_constants("./nonexistent_versioned_constants.json".as_ref()).unwrap_err(),
             ParseVersionedConstantsError::Io(err) => assert_eq!(err.kind(), std::io::ErrorKind::NotFound)
         );
     }
@@ -1060,22 +1058,23 @@ mod tests {
     #[test]
     fn parse_versioned_constants_fails_on_parse_error() {
         assert_matches!(
-            super::parse_versioned_constants("fixtures/invalid_versioned_constants.json".into())
+            super::parse_versioned_constants("fixtures/invalid_versioned_constants.json".as_ref())
                 .unwrap_err(),
-            ParseVersionedConstantsError::Parse(_)
+            ParseVersionedConstantsError::ParseMap(_)
         )
     }
 
     #[test]
     fn parse_versioned_constants_legacy() {
         super::parse_versioned_constants(
-            "../executor/resources/versioned_constants_0_13_1_1.json".into(),
+            "fixtures/blockifier_versioned_constants_0_13_1_1.json".as_ref(),
         )
         .unwrap();
     }
 
     #[test]
     fn parse_versioned_constants_success() {
-        super::parse_versioned_constants("fixtures/multi_versioned_constants.json".into()).unwrap();
+        super::parse_versioned_constants("fixtures/multi_versioned_constants.json".as_ref())
+            .unwrap();
     }
 }